Add jointrip event and update trip
Join trip event is accessable via url:
'/api/events/jointrip/<user_id>/<trip_id>'
It appends the user to the trips participants list.

Also added an update trip method that checks if the trips
date is today, if yes - updates the trips boolean field
isTripToday to True.

this method is called on GET requests for trip or trips.